Size the LTC frame-overflow sentinel to the target's own size_t
The overflow sentinel was written as a 64-bit constant, so the agent could
not be built for any 32-bit target: size_t is 32 bits on armv7 and the
constant does not fit.

    internal/agent/audio/ltcgen/ltcgen_cgo.go:144:19: constant
    18446744073709551615 overflows _Ctype_size_t

Use all-ones of whatever size_t is on the target, which matches the C
helper's own (size_t)-1 return and is the identical value on 64-bit.

The release workflow's note explaining why there is no armv7 job quoted
that compile error as the reason. It is corrected: the build now
succeeds, and what remains is packaging, because go env GOOS GOARCH names
that tarball linux_arm rather than linux_armv7 and the release job's
filenames assume the pattern the other two platforms share.

+++ internal/agent/audio/ltcgen/ltcgen_cgo.go
@@ -141,7 +141,9 @@ func (e *Encoder) NextFrame() ([]byte, pkgaudio.LTCTimecode, error) {
 	tc := pkgaudio.LTCTimecode(fmt.Sprintf("%02d:%02d:%02d:%02d", int(t.hours), int(t.mins), int(t.secs), int(t.frame)))
 
 	n := C.showmesh_ltc_frame_s16le(e.enc, (*C.int16_t)(unsafe.Pointer(&e.buf[0])), C.size_t(len(e.buf)))
-	if n == C.size_t(^uint64(0)) {
+	// The sentinel is whatever all-ones is on THIS target, matching the C
+	// helper's own (size_t)-1: size_t is 32 bits on armv7 and 64 on amd64.
+	if n == ^C.size_t(0) {
 		return nil, "", fmt.Errorf("ltcgen: libltc frame exceeded the %d-sample scratch buffer", len(e.buf))
 	}
 	if n == 0 {
